refactor(networkoperatorplugin): split discovery into helm-free sub-package
# Why

The library entry points `Discover` and `ParseClusterConfig` (added in
PR #102) live in `pkg/networkoperatorplugin`, which also hosts the
deploy/validate path (`helm.go`, `validate.go`, `deploy.go`). Because Go
compiles every `.go` file in a package, importing the library API drags
`helm.sh/helm/v3` and its transitive deps — including the MPL-2.0-licensed
`hashicorp/go-multierror` — into the consumer's vendor tree.

NVIDIA AICR's network-topology snapshotter needs the discovery library
but cannot accept MPL-2.0 in its license check, and pays a ~600-file
vendor diff just to render a hardware-discovered Measurement. This
commit splits the library entry points into a Helm-free sub-package so
external consumers stop paying for the deploy path's deps.

# What changes

## New package pkg/networkoperatorplugin/discovery

The library API moves here:

  - `Discover(ctx, kubeClient, restConfig, cfg, opts...)` — cluster
    discovery, populates cfg in place. Signature change: cfg is now a
    required positional argument; passing nil returns an error. The old
    `WithBaseConfig(*config.LaunchKitConfig)` option is removed —
    callers pass the cfg directly. The embedded default is still
    available via `config.DefaultLaunchKitConfig()` for callers who
    want l8k's baked-in defaults (currently NO release 26.4).
  - `ParseClusterConfig(io.Reader)` — unchanged behaviour, now lives
    here next to Discover.
  - New `WithRelease(string)` option pins the Network Operator release
    line by resolving the embedded catalog and writing the resolved
    version / repositories / helm-repo URL / DOCA driver version into
    cfg before discovery runs. Removes the need for library callers to
    construct those fields manually when overriding the default.
  - Existing options (`WithLogger`, `WithNodeSelector`,
    `WithKeepNamespace`, `WithCollapseNicRails`) carry over.
  - `Options` struct is now exported so the parent package's plugin
    method can populate it directly without going through the
    functional-options API.

`DiscoverClusterConfig` converts from a method on
`*NetworkOperatorPlugin` (in the old `discovery.go`) to a free function
`discovery.DiscoverClusterConfig` taking `(ctx, client, restConfig, cfg,
opts)`. The parent package keeps the `Plugin` interface method via a
five-line wrapper:

```go
func (p *NetworkOperatorPlugin) DiscoverClusterConfig(ctx context.Context,
    c client.Client, cfg *config.LaunchKitConfig) error {
    return discovery.DiscoverClusterConfig(ctx, c, p.RESTConfig, cfg,
        discovery.Options{
            NodeSelector:     p.NodeSelector,
            KeepNamespace:    p.KeepNamespace,
            CollapseNicRails: p.CollapseNicRails,
        })
}
```

`pkg/app/discover.go` is unchanged — it still calls
`plugin.DiscoverClusterConfig(...)` via the Plugin interface.

Files moved into the sub-package: `discovery.go` (→ `discover.go`),
`library.go`, `gputopology.go`, their `_test.go` siblings, the embedded
`ns-product-ids` file, plus `collapse_test.go` and `TestParseModuleList`
from `templates_test.go` (these exercised discovery internals).

## New package pkg/networkoperatorplugin/releases

Embedded release catalog (`releases.yaml`) and `LookupRelease /
SupportedReleases` move out of the parent package so the discovery
sub-package can read it without dragging the parent's Helm deps. Both
the parent's `ApplyNetworkOperatorRelease` (CLI/launcher path) and the
new `discovery.WithRelease` option consume it.

Callers updated: `pkg/cmd/{root,discover,generate,schema}.go`,
`pkg/networkoperatorplugin/networkoperator.go`, validate.go. The
top-level `releases.yaml` symlink is repointed at the new location.

## New package pkg/networkoperatorplugin/internal/pfutil

PF / NIC slice helpers (`FilterEastWestPFs`, `CollapsePFsToOnePerNIC`,
`IsDualPortModel`, `PciBusDevicePrefix`) move here so the discovery
sub-package and the parent's templates/render-plan paths can share one
implementation. Internal scope keeps them private to the l8k module.

Tests for `IsDualPortModel` and `CollapsePFsToOnePerNIC` move with the
helpers. Tests that exercised the wider `buildClusterConfig` flow
(`TestBuildClusterConfig_CollapseNicRails`,
`TestBuildClusterConfig_DropsNorthSouthOnlyGroups`) live with discovery
now since that function is in the discovery sub-package.

## Parent package surface unchanged

`pkg/networkoperatorplugin/{helm,validate,deploy,templates,
networkoperator,render_plan,scopes,workload,helper}.go` keep their
package and exported symbols. `NetworkOperatorPlugin` continues to
satisfy `plugin.Plugin`. The `helm.sh/helm/v3` deps the deploy path
depends on stay where they are — the goal of this PR is to keep the
library API helm-free, not to overhaul deploy.

# Verification

`go list -deps github.com/nvidia/k8s-launch-kit/pkg/networkoperatorplugin/discovery`:

  - Before (when discovery was in the parent package): 977 transitive
    deps including `helm.sh/helm/v3/pkg/action`, `helm.sh/helm/v3/pkg/kube`,
    `hashicorp/go-multierror` (MPL-2.0), `containerd/*`, `lib/pq`,
    `jmoiron/sqlx`, `lann/builder`, `gosuri/uitable`, `oras.land/oras-go`,
    `Masterminds/squirrel`, `chai2010/gettext-go`, …
  - After: 622 transitive deps. Zero `helm.sh/*`, zero `hashicorp/*`,
    zero `containerd/*`, zero SQL drivers, zero OCI client. The only
    `moby/*` entry (`spdystream`) comes via `k8s.io/client-go`'s pod-exec
    machinery — unrelated to Helm.

That's 355 fewer files in any vendor tree that imports the discovery
package — and crucially, the MPL-2.0 transitive dep is gone, so
license-check gates that disallow MPL-2.0 now pass.

`go test ./pkg/networkoperatorplugin/...` passes including the moved
discover/library/collapse/parsing tests. The pre-existing
`pkg/presets/TestGetPresetsDir_*` environmental failures are unaffected.

`golangci-lint run ./pkg/networkoperatorplugin/...` returns 0 issues.

# Library consumer API change

```go
// before:
import l8kplugin "github.com/nvidia/k8s-launch-kit/pkg/networkoperatorplugin"
cfg, err := l8kplugin.Discover(ctx, kc, restCfg,
    l8kplugin.WithBaseConfig(myCfg))

// after:
import (
    l8kconfig "github.com/nvidia/k8s-launch-kit/pkg/config"
    l8kdisc   "github.com/nvidia/k8s-launch-kit/pkg/networkoperatorplugin/discovery"
)
cfg, err := l8kconfig.DefaultLaunchKitConfig()
cfg, err = l8kdisc.Discover(ctx, kc, restCfg, cfg)
// or l8kdisc.Discover(ctx, kc, restCfg, myCfg) to use a caller-supplied config
// or l8kdisc.Discover(..., cfg, l8kdisc.WithRelease("26.1")) to override the release
```

Aicr's vendored l8k version will need its pseudo-version pin bumped and
its imports updated as a follow-up.
