+++ pkg/networkoperatorplugin/discovery/discover_test.go
@@ -14,7 +14,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package networkoperatorplugin
+package discovery
 
 import (
 	"context"
