+++ pkg/networkoperatorplugin/discovery/ns-product-ids
