+++ pkg/networkoperatorplugin/releases/releases.yaml
