+++ pkg/networkoperatorplugin/releases/releases_test.go
@@ -14,7 +14,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package networkoperatorplugin
+package releases
 
 import (
 	"testing"
