+++ releases.yaml
@@ -1,1 +1,1 @@
-pkg/networkoperatorplugin/releases.yaml
+pkg/networkoperatorplugin/releases/releases.yaml
